requirements.txt
requests
django
pytest
pytest-django
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-filter
psycopg2-binary